backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
anthropic==0.40.0
python-dotenv==1.0.1
pyyaml==6.0.2
twilio==9.3.7
httpx==0.27.2
websockets==13.1
python-multipart==0.0.12
icalendar==6.1.0
alembic==1.13.3
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
aiofiles==23.2.1
openpyxl==3.1.5
